requirements.txt
fastapi==0.104.1
uvicorn==0.34.0
SQLAlchemy==2.0.40
python-dotenv==1.0.0
psycopg2-binary==2.9.10 # Driver para PostgreSQL
alembic==1.12.1         # Para migraciones de base de datos
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
python-multipart>=0.0.5